llm-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
llama-cpp-python==0.2.90
python-multipart==0.0.20
requests==2.32.3
